using integral values rather than floats: this is safer for comparisons

+++ App/cogito/src/test/AlphaBetaAgent.java
@@ -22,11 +22,11 @@ public class AlphaBetaAgent extends MiniMaxAgent
   /* OVERRIDES */
 
   @Override
-  protected float evaluate(Rules rules, BoardMatrix board, Player current, 
+  protected int evaluate(Rules rules, BoardMatrix board, Player current, 
                           int depth)
   {
     return evaluate_ab(rules, board, current, depth, 
-            Float.MIN_VALUE, Float.MAX_VALUE);
+            Integer.MIN_VALUE, Integer.MAX_VALUE);
   }
   
   @Override
@@ -39,14 +39,14 @@ public class AlphaBetaAgent extends MiniMaxAgent
   
   /* SUBROUTINES */
   
-  private float evaluate_ab(Rules rules, BoardMatrix board, Player current, 
-                          int depth, float alpha, float beta)
+  private int evaluate_ab(Rules rules, BoardMatrix board, Player current, 
+                          int depth, int alpha, int beta)
   {
     // maximise my own gain, not the other player's
     Player me = getPlayer();
 
     // Check whether this is a leaf-node (victory, defeat, draw)
-    Float value = value(rules, board, me, depth);
+    Integer value = value(rules, board, me, depth);
     if(value != null)
       return value;
 
@@ -65,12 +65,12 @@ public class AlphaBetaAgent extends MiniMaxAgent
             : evaluate_min_ab(rules, children, current, depth, alpha, beta);
   }
 
-  protected float evaluate_min_ab(Rules rules, List<BoardMatrix> children,
-                                Player current, int depth, float alpha, float beta)
+  protected int evaluate_min_ab(Rules rules, List<BoardMatrix> children,
+                                Player current, int depth, int alpha, int beta)
   {
       for(BoardMatrix child : children)
       {
-        float beta_prime = evaluate_ab(rules, child, Game.otherPlayer(current), 
+        int beta_prime = evaluate_ab(rules, child, Game.otherPlayer(current), 
                                       depth +1, alpha, beta);
         beta = Math.min(beta, beta_prime);
         if(beta < alpha)
@@ -79,12 +79,12 @@ public class AlphaBetaAgent extends MiniMaxAgent
       return beta;
   }
 
-  protected float evaluate_max_ab(Rules rules, List<BoardMatrix> children,
-                              Player current, int depth, float alpha, float beta)
+  protected int evaluate_max_ab(Rules rules, List<BoardMatrix> children,
+                              Player current, int depth, int alpha, int beta)
   {
       for(BoardMatrix child : children)
       {
-        float alpha_prime = evaluate_ab(rules, child, Game.otherPlayer(current), 
+        int alpha_prime = evaluate_ab(rules, child, Game.otherPlayer(current), 
                                       depth +1, alpha, beta);
         alpha = Math.max(alpha, alpha_prime);
         if(alpha >= beta)
